backend/requirements.txt
alembic==1.10.4
aniso8601==9.0.1
astroid==2.15.3
click==8.1.3
colorama==0.4.6
coverage==7.2.2
dill==0.3.6
Flask==2.2.3
Flask-Migrate==4.0.4
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
importlib-metadata==6.5.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
Mako==1.2.4
MarkupSafe==2.1.2
mccabe==0.7.0
platformdirs==3.2.0
python-dotenv==1.0.0
pytz==2023.3
regex==2023.5.4
six==1.16.0
SQLAlchemy==2.0.9
tomlkit==0.11.7
typing_extensions==4.5.0
Werkzeug==2.2.3
wrapt==1.15.0
zipp==3.15.0
